requirements.txt
Flask==3.1.0
openai==1.72.0
python-dotenv==1.1.0
textblob==0.19.0
transformers==4.51.2
